Refactor pipeline imports in main scripts and README for improved clarity and consistency
