+++ pgse/__init__.py
@@ -1,0 +1,5 @@
+from pgse.pipeline.pgse_pipeline import Pipeline as TrainingPipeline
+from pgse.pipeline.pgse_inference_pipeline import Pipeline as InferencePipeline
+from pgse.pipeline.regular_pipline import Pipeline as PureXGBPipeline
+
+__all__ = ['TrainingPipeline', 'InferencePipeline', 'PureXGBPipeline']
